// packages/astro-loader-github-releases/src/graphql/gen/operations.ts
export type GetReleaseByIdQuery = {
  node?:
    | {
        id: string
        url: string
        name?: string | null
        tagName: string
        description?: string | null
        descriptionHTML?: string | null
        isDraft: boolean
        isLatest: boolean
        isPrerelease: boolean
        createdAt: string
        publishedAt?: string | null
        repository: {
          name: string
          nameWithOwner: string
          url: string
          stargazerCount: number
          isInOrganization: boolean
        }
      }
    | {}
    | null
}